refactor: GameProcess를 GameController, GameService로 분리
