+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.Controller.GameController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        GameProcess gameProcess = new GameProcess();
-        gameProcess.startGame();
+        GameController gameController = new GameController();
+        gameController.startGame();
     }
 }
