+++ src/main/java/racingcar/Service/GameService.java
@@ -1,27 +1,16 @@
-package racingcar;
+package racingcar.Service;
 
-import camp.nextstep.edu.missionutils.Console;
 import camp.nextstep.edu.missionutils.Randoms;
+import racingcar.Car;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-public class GameProcess {
+public class GameService {
     private static final int MIN_NUMBER = 0;
     private static final int MAX_NUMBER = 9;
     private static final int FORWARD_NUMBER = 4;
-
-    public void startGame() {
-        System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
-        String carName = Console.readLine();
-        List<String> carList = parseCarName(carName);
-
-        System.out.println("시도할 회수는 몇회인가요?");
-        String countTrial = Console.readLine();
-
-        System.out.println(carList);
-    }
 
     // 쉼표(,)를 기준으로 입력된 자동차 이름 분해 후 List에 저장
     public List<String> parseCarName(String input) {
